Combine all import functions into the Importer.
import_legacy_csv is the main routine that the Rake
task will call.

The transmute function serves as an example of various CSV
transformations that might be made when importing.

The current selection of transformations support the current
club I’m working with. The functions and tests are somewhat
over-engineered and over-tested given the possibility that
the import will be run only once. But experience says that the
import will always need to be run multiple times – for bugs in
the process, or more likely, bad data in the imported file.