Refer to class method when calling auth.
This method was not recognized after I moved it to a Concern.
Though I did test the change and thought it worked – no explanation.

This guide was followed for creating the Concern:
https://medium.com/weareevermore/how-to-add-http-basic-authentication-to-your-rails-application-e4e4d5b958d9

Another user (without a Concern) reported the same error, and proposed
the fix I used:
https://github.com/thoughtbot/administrate/issues/1092

Other suggestions for the error said to Include the HTTP methods that cover
this area – adding the Include to the Concern didn’t help.